[feature] WidgetCard: create card template and migrate usage/projected consumption
