+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -54,10 +54,12 @@
     <string name="usage_estimated_cost">estimated</string>
     <string name="usage_reference_cost">reference*</string>
     <string name="usage_estimated_daily">Daily average: %1$s kWh / £%2$s</string>
-    <string name="usage_annual_projection">Projected annual consumption</string>
+    <string name="usage_projected_consumption">Projected Consumption</string>
     <string name="usage_demo_introduction">Demo Mode: The consumption figures shown are randomly generated.</string>
     <string name="usage_consumption">Consumption</string>
     <string name="usage_standing_charge">Standing Charge</string>
+    <string name="usage_kwh_year">kWh / year</string>
+    <string name="usage_kwh_month">(%1$s kWh / month)</string>
 
     <!-- Agile -->
     <string name="agile_demo_introduction">Demo Mode: The Agile Tariff shown for Retail Region A may not be available to you.</string>
